Ported to new wreport
